CSV generator

+++ Tarefa 2/src/setDistribution.hpp
@@ -8,18 +8,18 @@
 
 using namespace std;
 
-void equalSplit(unsigned int N, unsigned int M, unsigned int* elements, unsigned int* threadSize) {
+void equalSplit(long long N, long long M, long long* elements, long long* threadSize) {
     // Evaluate the number of elements per thread
-    unsigned int elementsPerThread = N / M;
-    unsigned int remainingElements = N % M;
+    long long elementsPerThread = N / M;
+    long long remainingElements = N % M;
 
     // Adjust the threadSize array for the remaining elements, if any
-    for (unsigned int i = 0; i < M; i++) {
+    for (long long i = 0; i < M; i++) {
         threadSize[i] = elementsPerThread + (i < remainingElements ? 1 : 0);
     }
 }
 
-void shuffleAndSplit(unsigned int N, unsigned int M, unsigned int* elements, unsigned int* threadSize) {
+void shuffleAndSplit(long long N, long long M, long long* elements, long long* threadSize) {
     // Shuffle the elements
     random_device rd;
     mt19937 g(rd()); // Mersenne Twister engine for good randomness
@@ -29,7 +29,7 @@ void shuffleAndSplit(unsigned int N, unsigned int M, unsigned int* elements, uns
     equalSplit(N, M, elements, threadSize);
 }
 
-void cardDistribution(unsigned int N, unsigned int M, unsigned int* elements, unsigned int* threadSize) {
+void cardDistribution(long long N, long long M, long long* elements, long long* threadSize) {
     // Call equalSplit the evaluate the number of elements per thread
     equalSplit(N, M, elements, threadSize);
 
@@ -37,15 +37,15 @@ void cardDistribution(unsigned int N, unsigned int M, unsigned int* elements, un
     fill(elements, elements + N, 0);
 
     // Distribute the numbers across elements array in the desired order
-    unsigned int currentThreadIndex = 0;
-    unsigned int currentThreadOffsets[M] = {0};
+    long long currentThreadIndex = 0;
+    long long currentThreadOffsets[M] = {0};
 
     // Fill the currentThreadOffsets with the start index based on the threadSize
-    for (unsigned int i = 1; i < M; i++) {
+    for (long long i = 1; i < M; i++) {
         currentThreadOffsets[i] = currentThreadOffsets[i - 1] + threadSize[i - 1];
     }
 
-    for (unsigned int i = 0; i < N; i++) {
+    for (long long i = 0; i < N; i++) {
         // Assign the current element to the current thread
         elements[currentThreadOffsets[currentThreadIndex]] = i + 1;
 
@@ -55,78 +55,78 @@ void cardDistribution(unsigned int N, unsigned int M, unsigned int* elements, un
     }
 }
 
-void splitAndShuffle(unsigned int N, unsigned int M, unsigned int* elements, unsigned int* threadSize) {
+void splitAndShuffle(long long N, long long M, long long* elements, long long* threadSize) {
 
     // Divide the orderedElements array into parts of size M and shuffle each part
-    unsigned int elementsPerThread = N / M;
-    unsigned int remainingElements = N % M;
+    long long elementsPerThread = N / M;
+    long long remainingElements = N % M;
 
     // Adjust the threadSize array for the remaining elements, if any
-    for (unsigned int i = 0; i < M; i++) {
+    for (long long i = 0; i < M; i++) {
         threadSize[i] = elementsPerThread + (i < remainingElements ? 1 : 0);
     }
 
     // Generate a new array with the ordered elements
-    unsigned int orderedElements[N] = {0};
-    for (unsigned int i = 0; i < N; i++) {
+    long long orderedElements[N] = {0};
+    for (long long i = 0; i < N; i++) {
         orderedElements[i] = i + 1;
     }
     // Shuffle each block of size M
     random_device rd;
     mt19937 g(rd()); // Mersenne Twister engine for good randomness
 
-    for (unsigned int i = 0; i < elementsPerThread; i++) {
+    for (long long i = 0; i < elementsPerThread; i++) {
         shuffle(orderedElements + i * M, orderedElements + (i + 1) * M, g);
     }
 
     // Shuffle the remaining elements
     shuffle(orderedElements + elementsPerThread * M, orderedElements + elementsPerThread * M + remainingElements, g);
 
-    unsigned int currentThreadIndex = 0;
-    unsigned int currentThreadOffsets[M] = {0};
+    long long currentThreadIndex = 0;
+    long long currentThreadOffsets[M] = {0};
 
     // Fill the currentThreadOffsets with the start index based on the threadSize
-    for (unsigned int i = 1; i < M; i++) {
+    for (long long i = 1; i < M; i++) {
         currentThreadOffsets[i] = currentThreadOffsets[i - 1] + threadSize[i - 1];
     }
 
     // Assign the shuffled elements to the elements array
     // Each threadSize will contain one element of each block of size M
     // The remaining elements will be assigned to the first threads
-    for (unsigned int i = 0; i < M; i++) {
-        for (unsigned int j = 0; j < threadSize[i]; j++) {
+    for (long long i = 0; i < M; i++) {
+        for (long long j = 0; j < threadSize[i]; j++) {
             elements[currentThreadOffsets[i] + j] = orderedElements[i + j * M];
         }
     }
 }
 
-void workloadBalance(unsigned int N, unsigned int M, unsigned int* elements, unsigned int* threadSize) {
+void workloadBalance(long long N, long long M, long long* elements, long long* threadSize) {
 
     // Calculate the workload for each element and the total workload using squares
-    unsigned int workload[N] = {0};
-    unsigned int totalWorkload = 0;
+    long long workload[N] = {0};
+    long long totalWorkload = 0;
 
-    for (unsigned int i = 0; i < sqrt(N); i++) {
+    for (long long i = 0; i < sqrt(N); i++) {
         // Select the range for the current square workload
-        unsigned int start = i * i;
-        unsigned int end = (i + 1) * (i + 1) - 1;
+        long long start = i * i;
+        long long end = (i + 1) * (i + 1) - 1;
         end = min(end, N); // Ensure we don't go beyond N
 
         // Evaluate the workload for each element with this workload
-        for (unsigned int j = start; j <= end; j++) {
+        for (long long j = start; j <= end; j++) {
             totalWorkload += i;
             workload[j] = i;
         }
     }
 
     // Calculate the target workload per thread
-    unsigned int targetWorkload = totalWorkload / M;
+    long long targetWorkload = totalWorkload / M;
 
     // Calculate the thread size
-    unsigned int currentThread = 0;
-    unsigned int currentWorkload = 0;
+    long long currentThread = 0;
+    long long currentWorkload = 0;
 
-    for (unsigned int i = 0; i < N; i++) {
+    for (long long i = 0; i < N; i++) {
         currentWorkload += workload[i];
         threadSize[currentThread]++;
